requirements.txt
streamlit==1.52.2
pandas==2.3.3
openpyxl==3.1.5
google-cloud-bigquery==3.39.0
sqlalchemy==2.0.45
sqlalchemy-bigquery==1.16.0
db-dtypes==1.5.0
pyarrow==22.0.0
plotly